requirements.txt
Django==4.2.10
psycopg2==2.9.9
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-storages==1.14.2
django-debug-toolbar==4.2.0
python-dotenv==1.0.0
djangorestframework==3.14.0
django-filter==23.5
django-allauth==0.59.0
django-taggit==5.0.1
minio==7.2.0
whitenoise==6.5.0
boto3